move resizable map into it’s own component
